better docs and simplify HTTPSUrl validator
