+++ examples/resources/hcp_notifications_webhook/resource.tf
@@ -1,7 +1,6 @@
 resource "hcp_notifications_webhook" "example" {
   name        = "example-webhook"
   description = "Notify for all of the events for all Packer artifact versions existing in the project."
-  enabled     = true
 
   config = {
     url = "https://example.com"
