--- internal/hcpvalidator/https_url_test.go
@@ -1,64 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package hcpvalidator_test
-
-import (
-	"context"
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-framework/path"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-provider-hcp/internal/hcpvalidator"
-)
-
-func TestHTTPSUrlValidator(t *testing.T) {
-	t.Parallel()
-
-	type testCase struct {
-		val         types.String
-		expectError bool
-	}
-	tests := map[string]testCase{
-		"unknown String": {
-			val: types.StringUnknown(),
-		},
-		"null String": {
-			val: types.StringNull(),
-		},
-		"valid http url": {
-			val:         types.StringValue("http://url.com"),
-			expectError: true,
-		},
-		"valid https url": {
-			val: types.StringValue("https://url.com"),
-		},
-		"invalid url": {
-			val:         types.StringValue("invalid"),
-			expectError: true,
-		},
-	}
-
-	for name, test := range tests {
-		name, test := name, test
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			request := validator.StringRequest{
-				Path:           path.Root("test"),
-				PathExpression: path.MatchRoot("test"),
-				ConfigValue:    test.val,
-			}
-			response := validator.StringResponse{}
-			hcpvalidator.HTTPSUrl().ValidateString(context.TODO(), request, &response)
-
-			if !response.Diagnostics.HasError() && test.expectError {
-				t.Fatal("expected error, got no error")
-			}
-
-			if response.Diagnostics.HasError() && !test.expectError {
-				t.Fatalf("got unexpected error: %s", response.Diagnostics)
-			}
-		})
-	}
-}
